+++ packages/@magic-sdk/provider/jest.config.ts
@@ -1,1 +1,12 @@
-export { default } from '../../../jest.config';
+import baseJestConfig from '../../../jest.config';
+import type { Config } from '@jest/types';
+
+const config: Config.InitialOptions = {
+  ...baseJestConfig,
+  transform: {
+    '^.+\\.(js|jsx)$': 'babel-jest',
+    '\\.(ts|tsx)$': 'ts-jest',
+  },
+};
+
+export default config;