+++ packages/@magic-sdk/provider/src/core/sdk-environment.ts
@@ -1,8 +1,7 @@
-import localForage from 'localforage';
-import { ViewController } from './view-controller';
-import { PayloadTransport } from './payload-transport';
-import { SDKBase } from './sdk';
-import { WithExtensions } from '../modules/base-extension';
+import type localForage from 'localforage';
+import type { ViewController } from './view-controller';
+import type { SDKBase } from './sdk';
+import type { WithExtensions } from '../modules/base-extension';
 
 type ConstructorOf<C> = { new (...args: any[]): C };
 
@@ -16,7 +15,6 @@ export interface SDKEnvironment {
   platform: 'web' | 'react-native';
   defaultEndpoint: string;
   ViewController: ConstructorOf<ViewController>;
-  PayloadTransport: ConstructorOf<PayloadTransport>;
   configureStorage: () => Promise<typeof localForage>;
 }