+++ packages/@magic-sdk/provider/src/util/index.ts
@@ -1,7 +1,10 @@
+export * from './base64-json';
+export * from './events';
 export * from './get-payload-id';
 export * from './promise-tools';
-export * from './base64-json';
-export * from './type-guards';
-export * from './events';
-export * from './url';
+export * as semver from './semver';
 export * as storage from './storage';
+export * from './type-guards';
+export * from './url';
+export * from './uuid';
+export * from './web-crypto';