+++ scripts/utils/print-separator.ts
@@ -1,0 +1,9 @@
+import chalk from 'chalk';
+
+/**
+ * Prints a labeled separator to create a nice visual
+ * distinction between sections of script output.
+ */
+export function printSeparator(label: string) {
+  console.log(chalk`\n{dim ❮❮❮} ${label} {dim ❯❯❯}\n`);
+}