+++ packages/@magic-sdk/react-native/CHANGELOG.md
@@ -1,3 +1,39 @@
+# v6.0.2 (Tue Sep 14 2021)
+
+#### 🐛 Bug Fix
+
+- Import regeneratorRuntime in Magic JS (non-CDN version) [#210](https://github.com/magiclabs/magic-js/pull/210) ([@smithki](https://github.com/smithki))
+
+#### Authors: 1
+
+- Ian K Smith ([@smithki](https://github.com/smithki))
+
+---
+
+# v6.0.1 (Tue Sep 14 2021)
+
+#### 🐛 Bug Fix
+
+- Fix SemVer cyclic dependency issues with some hacks [#209](https://github.com/magiclabs/magic-js/pull/209) ([@smithki](https://github.com/smithki))
+
+#### Authors: 1
+
+- Ian K Smith ([@smithki](https://github.com/smithki))
+
+---
+
+# v6.0.0 (Tue Sep 14 2021)
+
+#### 💥 Breaking Change
+
+- v6.0.0 [#208](https://github.com/magiclabs/magic-js/pull/208) ([@smithki](https://github.com/smithki))
+
+#### Authors: 1
+
+- Ian K Smith ([@smithki](https://github.com/smithki))
+
+---
+
 # v4.4.2 (Mon Aug 16 2021)
 
 #### 🐛 Bug Fix