+++ packages/@magic-sdk/types/CHANGELOG.md
@@ -1,3 +1,27 @@
+# v5.0.1 (Tue Sep 14 2021)
+
+#### 🐛 Bug Fix
+
+- Import regeneratorRuntime in Magic JS (non-CDN version) [#210](https://github.com/magiclabs/magic-js/pull/210) ([@smithki](https://github.com/smithki))
+
+#### Authors: 1
+
+- Ian K Smith ([@smithki](https://github.com/smithki))
+
+---
+
+# v5.0.0 (Tue Sep 14 2021)
+
+#### 💥 Breaking Change
+
+- v6.0.0 [#208](https://github.com/magiclabs/magic-js/pull/208) ([@smithki](https://github.com/smithki))
+
+#### Authors: 1
+
+- Ian K Smith ([@smithki](https://github.com/smithki))
+
+---
+
 # v4.0.0 (Tue Aug 31 2021)
 
 #### 💥 Breaking Change